# Omit FETCH_THIS in closures