@mvorisek 2024-05-09
This PR seems to be related with https://github.com/php/php-src/issues/10782, can it fix it?