@iluuu1994 2024-05-09
@mvorisek I don't think this will fix the mentioned problem. It is related to run_time_cache per closure, this only slightly tweaks the opcodes of the closure.