@staabm 2024-05-10
Does this change improve closure performance?