@iluuu1994 2024-05-10
It seems so.

```php
class C {
    public function test() {
        $c = function () {
            for ($i = 0; $i < 100000000; $i++) {
                $this->foo();
            }
        };
        $c();
    }

    private function foo() {}
}

$c = new C();
$c->test();
```

Before:

```
~/Developer/php-src @d2a9edfe ?1 ❯ hyperfine 'taskset --cpu-list 0 php-dev -d opcache.enable_cli=1 test.php'                                                 8s
Benchmark 1: taskset --cpu-list 0 php-dev -d opcache.enable_cli=1 test.php
  Time (mean ± σ):     795.8 ms ±   6.9 ms    [User: 790.8 ms, System: 2.7 ms]
  Range (min … max):   779.1 ms … 805.2 ms    10 runs
```

After:

```
~/Developer/php-src omit-fetch-this-in-closures ?1 ❯ hyperfine 'taskset --cpu-list 0 php-dev -d opcache.enable_cli=1 test.php'                               4s
Benchmark 1: taskset --cpu-list 0 php-dev -d opcache.enable_cli=1 test.php
  Time (mean ± σ):     570.1 ms ±  15.6 ms    [User: 565.7 ms, System: 2.8 ms]
  Range (min … max):   557.7 ms … 609.9 ms    10 runs
```

But this is highly synthetic, so take it with a grain of salt. I'm not sure whether it improves real-life performance.