@dstogov 2024-05-13
I like the patch, and in general everything looks right.
I tested it with PHP-Parser benchmark and it shows slight improvement with function JIT and without JIT.
However it also shows degradation with tracing JIT.
For some reason, PHP starts to generate bigger and slower code.

Can you investigate this please. (I suppose, JIT have some CLOSURE specific optimization that stopped work after your patch).