@dstogov 2024-05-13
This demonstrates the worse JIT code generated for FETCH_DIM_R.

```php
<?php
class Foo {
	public $a = 42;
	public function test() {
		return function() {
			return $this->a;
		};
	}
}
$obj = new Foo();
$fn = $obj->test();
for ($i = 0; $i < 100; $i++) {
	$fn();
}
?>
```

```
sapi/cli/php -d opcache.jit_debug=0x1ff201 fetch_this.php 2>err1.log
sapi/cli/php.patched -d opcache.jit_debug=0x1ff201 fetch_this.php 2>err2.log
diff -u err1.log err2.log
```

The degradation is caused by call to ``zend_jit_fetch_obj()``. Without the patch it gets ``trace_ce`` recorded by the tracer and generates class guards and optimized code. With the patch ``trace_ce`` is NULL and the slow code is generated.

It's possible that we have similar degradation for other related instructions. (ASSIGN_OBJ, INIT_METHOD_CALL, etc)

It looks like the following extension to the tracer fixes all degradation. May be it should record $this class only for closures. (I'm not sure). 

```diff
diff --git a/ext/opcache/jit/zend_jit_vm_helpers.c b/ext/opcache/jit/zend_jit_vm_helpers.c
index 7df2cd5b902..4d055ac2ece 100644
--- a/ext/opcache/jit/zend_jit_vm_helpers.c
+++ b/ext/opcache/jit/zend_jit_vm_helpers.c
@@ -692,6 +692,25 @@ zend_jit_trace_stop ZEND_FASTCALL zend_jit_trace_execute(zend_execute_data *ex,
 				}
 			}
 			op1_type |= flags;
+		} else if (opline->op1_type == IS_UNUSED
+		 && (opline->opcode == ZEND_FETCH_OBJ_R
+		  || opline->opcode == ZEND_FETCH_OBJ_W
+		  || opline->opcode == ZEND_FETCH_OBJ_RW
+		  || opline->opcode == ZEND_FETCH_OBJ_IS
+		  || opline->opcode == ZEND_FETCH_OBJ_FUNC_ARG
+		  || opline->opcode == ZEND_FETCH_OBJ_UNSET
+		  || opline->opcode == ZEND_ASSIGN_OBJ
+		  || opline->opcode == ZEND_ASSIGN_OBJ_REF
+		  || opline->opcode == ZEND_ASSIGN_OBJ_OP
+		  || opline->opcode == ZEND_PRE_INC_OBJ
+		  || opline->opcode == ZEND_PRE_DEC_OBJ
+		  || opline->opcode == ZEND_POST_INC_OBJ
+		  || opline->opcode == ZEND_POST_DEC_OBJ
+		  || opline->opcode == ZEND_ISSET_ISEMPTY_PROP_OBJ
+		  || opline->opcode == ZEND_UNSET_OBJ
+		  || opline->opcode == ZEND_INIT_METHOD_CALL
+		  || opline->opcode == ZEND_CLONE)) {
+			ce1 = Z_OBJCE(EX(This));
 		}
 		if (opline->op2_type & (IS_TMP_VAR|IS_VAR|IS_CV)
 		 && opline->opcode != ZEND_INSTANCEOF
```

@iluuu1994 Please finalize this.